+++ src/pydiverse/common/util/computation_tracing.py
@@ -1,8 +1,5 @@
 # Copyright (c) QuantCo and pydiverse contributors 2025-2025
 # SPDX-License-Identifier: BSD-3-Clause
-
-from __future__ import annotations
-
 import dis
 import inspect
 from enum import Enum
@@ -26,7 +23,7 @@ class Operation(Enum):
 
 
 class ComputationTracer:
-    proxy_type: type[ComputationTracerProxy]
+    proxy_type: type["ComputationTracerProxy"]
 
     def __init__(self):
         self.trace = []
@@ -86,8 +83,7 @@ class ComputationTracer:
             from dask.base import tokenize
         except ModuleNotFoundError:
             raise ModuleNotFoundError(
-                "module dask is required to use @materialize(version=AUTO_VERSION)"
-                " or computation_tracing."
+                "module dask is required to use computation_tracing."
             ) from None
 
         return tokenize(self.trace)