+++ src/pydiverse/common/util/structlog.py
@@ -1,8 +1,5 @@
 # Copyright (c) QuantCo and pydiverse contributors 2025-2025
 # SPDX-License-Identifier: BSD-3-Clause
-
-from __future__ import annotations
-
 import logging
 import sys
 import textwrap
@@ -23,6 +20,7 @@ except ImportError:
 
     structlog.dev = dev
     structlog_installed = False
+    EventDict, WrappedLogger = None, None
 
 
 class StructlogHandler(logging.Handler):